+++ packages/react-composites/tests/browser/chat/app/fake-back-end/ThreadEventEmitter.ts
